+++ CHANGELOG.md
@@ -112,3 +112,4 @@
 - `Mgmg::IR`を追加．
 	- これを生成するための`String#ir`および`Enumerable#ir`を追加
 	- `String#search`および`Enumerable#search`において，内部的にこれを利用することで高速化．
+- `Mgmg.#find_lowerbound`, `Mgmg.#find_upperbound`を追加．